Remove HealthCheckController and rename GetLatestController to Api\V0 namespace
